feat: plugin.RUN is the entrypoint of the business logic
