+++ src/plugin/config_test.go
@@ -26,10 +26,12 @@ func TestFetchConfigFromEnvironment(t *testing.T) {
 	var config plugin.Config
 	fetcher := plugin.EnvironmentConfigFetcher{}
 
-	t.Setenv("BUILDKITE_PLUGIN_ECS_TASK_RUNNER_MESSAGE", "test-message")
+	t.Setenv("BUILDKITE_PLUGIN_ECS_TASK_RUNNER_PARAMETER_NAME", "test-parameter")
+	t.Setenv("BUILDKITE_PLUGIN_ECS_TASK_RUNNER_SCRIPT", "hello-world")
 
 	err := fetcher.Fetch(&config)
 
 	require.NoError(t, err, "fetch should not error")
-	assert.Equal(t, "test-message", config.Message, "fetched message should match environment")
+	assert.Equal(t, "test-parameter", config.ParameterName, "fetched message should match environment")
+	assert.Equal(t, "hello-world", config.Script, "fetched message should match environment")
 }
