--- src/plugin/example_test.go
@@ -1,25 +1,0 @@
-package plugin_test
-
-import (
-	"context"
-	"testing"
-
-	"ecs-task-runner/plugin"
-
-	"github.com/stretchr/testify/assert"
-)
-
-func TestDoesAnnotate(t *testing.T) {
-	agent := &AgentMock{}
-	fetcher := plugin.EnvironmentConfigFetcher{}
-	examplePlugin := plugin.ExamplePlugin{}
-	ctx := context.Background()
-	annotation := "test-message"
-
-	t.Setenv("BUILDKITE_PLUGIN_ECS_TASK_RUNNER_MESSAGE", annotation)
-	agent.Mock.On("Annotate", ctx, annotation, "info", "message").Return(nil)
-
-	err := examplePlugin.Run(ctx, fetcher, agent)
-
-	assert.NoError(t, err, "should not error")
-}
